fixing all compiler warnings

+++ RNMC/dispatcher.h
@@ -182,12 +182,12 @@ template <typename Solver>
 void Dispatcher<Solver>::record_simulation_history(HistoryPacket history_packet) {
     // TODO: break into subtransactions for super long simulations
     initial_state_database.exec("BEGIN");
-    for (int i = 0; i < history_packet.history.size(); i++) {
+    for (unsigned long int i = 0; i < history_packet.history.size(); i++) {
         trajectories_writer.insert(
 
             TrajectoriesSql {
                 .seed = (int) history_packet.seed,
-                .step = i,
+                .step = (int) i,
                 .reaction_id = history_packet.history[i].reaction,
                 .time = history_packet.history[i].time
             });
